Add minimal -devel packages as needed to build kernel modules
As discussed in the forum[1], package layering can work for building kernel
modules, such as the nvidia driver. However, there are some problems
with the -devel packages needed, as they need to match the exact version
in the base image, and once the version bumps in the yum repo the old
version isn't even available.

So, this adds the minimal set of -devel packages needed to build a a kmod.
This isn't enough to build a kernel module, but the rest needed can much
more easily be layered.

Given the current rawhide dependencies this adds the following packages:
 * elfutils-libelf-devel
 * glibc-devel
 * glibc-headers
 * kernel-devel
 * kernel-headers
 * libxcrypt-devel
 * zlib-devel

This adds about 100 megs to a 4 gigabyte image, which I think is a fair
compromise in order to be able to build kernel modules and support
the nvidia drivers.

[1] https://discussion.fedoraproject.org/t/supporting-the-nvidia-drivers-on-silverblue/849/6