Exclude python3-unbound
This package creates the unbound user and group which are not used by
any system daemon. Excluding it thus simplifies /etc/passwd & /etc/group
management.

This package is a recommended dependency pulled by python3-dnf which is
itself pulled by initial-setup-gui.

This impacts all the desktop environments that replaced
gnome-initial-setup with initial-setup-gui (KDE, XFCE & LXQt as of this
commit).